refactor benchmark metric ownership
