+++ benchmarks/eval/benchmark_tts_seedtts.py
@@ -119,12 +119,14 @@ from dataclasses import dataclass
 from benchmarks.benchmarker.runner import BenchmarkRunner, RunConfig
 from benchmarks.benchmarker.utils import wait_for_service
 from benchmarks.dataset.seedtts import load_seedtts_samples
-from benchmarks.metrics.performance import compute_speed_metrics
+from benchmarks.metrics.performance import (
+    build_speed_results,
+    compute_speed_metrics,
+    print_speed_summary,
+)
 from benchmarks.tasks.tts import (
     build_base_url,
-    build_speed_results,
     make_tts_send_fn,
-    print_speed_summary,
     run_seedtts_transcribe,
     save_generated_audio_metadata,
     save_speed_results,
