+++ benchmarks/tasks/tts.py
@@ -1,6 +1,5 @@
 # SPDX-License-Identifier: Apache-2.0
-"""TTS task utilities: voice-clone API clients, WER/ASR evaluation,
-HTTP send functions, and speed/WER result builders.
+"""TTS task utilities: voice-clone API clients, ASR evaluation, and HTTP send functions.
 
 Replaces tasks/tts_speed.py and tasks/voice_clone.py.
 """
@@ -24,7 +23,6 @@ from pathlib import Path
 from typing import Protocol
 
 import aiohttp
-import numpy as np
 import soundfile as sf
 import torch
 import transformers
@@ -43,12 +41,17 @@ from benchmarks.benchmarker.utils import (
     save_json_results,
 )
 from benchmarks.dataset.seedtts import SampleInput
+from benchmarks.metrics.performance import build_speed_results
+from benchmarks.metrics.wer import (
+    calculate_asr_speed_metrics,
+    calculate_wer_metrics,
+    print_asr_speed_summary,
+    print_wer_summary,
+)
 
 logger = logging.getLogger(__name__)
 
 TEXT_PREVIEW_LENGTH = 60
-SUMMARY_LABEL_WIDTH = 30
-SUMMARY_LINE_WIDTH = 60
 
 
 # ---------------------------------------------------------------------------
@@ -239,68 +242,6 @@ def transcribe_and_compute_wer(
     return output
 
 
-def calculate_wer_metrics(outputs: list[SampleOutput], lang: str) -> dict:
-    """Compute corpus-level WER metrics from per-sample outputs."""
-    successes = [o for o in outputs if o.is_success]
-    if not successes:
-        return {
-            "lang": lang,
-            "total_samples": len(outputs),
-            "evaluated": 0,
-            "skipped": len(outputs),
-            "wer_corpus": 0.0,
-            "wer_per_sample_mean": 0.0,
-            "wer_per_sample_median": 0.0,
-            "wer_per_sample_std": 0.0,
-            "wer_per_sample_p95": 0.0,
-            "wer_per_sample_max": 0.0,
-            "wer_below_50_corpus": 0.0,
-            "n_above_50_pct_wer": 0,
-            "pct_above_50_pct_wer": 0.0,
-            "latency_mean_s": 0.0,
-            "audio_duration_mean_s": 0.0,
-        }
-
-    total_errors = sum(o.substitutions + o.deletions + o.insertions for o in successes)
-    total_ref_words = sum(o.substitutions + o.deletions + o.hits for o in successes)
-    corpus_wer = total_errors / total_ref_words if total_ref_words > 0 else 0.0
-
-    wer_arr = np.array([o.wer for o in successes])
-    latencies = [o.latency_s for o in successes]
-    audio_durations = [o.audio_duration_s for o in successes if o.audio_duration_s > 0]
-
-    n_above_50 = int(np.sum(wer_arr > 0.5))
-    ok_samples = [o for o in successes if o.wer <= 0.5]
-    if ok_samples:
-        ok_errors = sum(
-            o.substitutions + o.deletions + o.insertions for o in ok_samples
-        )
-        ok_ref = sum(o.substitutions + o.deletions + o.hits for o in ok_samples)
-        wer_below_50_micro = ok_errors / ok_ref if ok_ref > 0 else 0.0
-    else:
-        wer_below_50_micro = 0.0
-
-    return {
-        "lang": lang,
-        "total_samples": len(outputs),
-        "evaluated": len(successes),
-        "skipped": len(outputs) - len(successes),
-        "wer_corpus": float(corpus_wer),
-        "wer_per_sample_mean": float(np.mean(wer_arr)),
-        "wer_per_sample_median": float(np.median(wer_arr)),
-        "wer_per_sample_std": float(np.std(wer_arr)),
-        "wer_per_sample_p95": float(np.percentile(wer_arr, 95)),
-        "wer_per_sample_max": float(np.max(wer_arr)),
-        "wer_below_50_corpus": float(wer_below_50_micro),
-        "n_above_50_pct_wer": n_above_50,
-        "pct_above_50_pct_wer": (n_above_50 / len(successes) * 100 if successes else 0),
-        "latency_mean_s": float(np.mean(latencies)),
-        "audio_duration_mean_s": (
-            float(np.mean(audio_durations)) if audio_durations else 0
-        ),
-    }
-
-
 def compute_text_audio_consistency(
     request_results: list[RequestResult],
     lang: str,
@@ -341,171 +282,6 @@ def compute_text_audio_consistency(
         for o in outputs
     ]
     return {"summary": calculate_wer_metrics(outputs, lang), "per_sample": per_sample}
-
-
-def print_wer_summary(
-    metrics: dict, model_name: str, generation_mode: str | None = None
-) -> None:
-    lw = SUMMARY_LABEL_WIDTH
-    w = SUMMARY_LINE_WIDTH
-    title = "TTS WER Benchmark Result"
-    if generation_mode:
-        title = f"TTS WER Benchmark Result ({generation_mode})"
-    print(f"\n{'=' * w}")
-    print(f"{title:^{w}}")
-    print(f"{'=' * w}")
-    print(f"  {'Model:':<{lw}} {model_name}")
-    if generation_mode:
-        print(f"  {'Generation mode:':<{lw}} {generation_mode}")
-    print(f"  {'Language:':<{lw}} {metrics.get('lang', 'N/A')}")
-    print(
-        f"  {'Evaluated / Total:':<{lw}} "
-        f"{metrics.get('evaluated', 0)}/{metrics.get('total_samples', 0)}"
-    )
-    print(f"  {'Skipped:':<{lw}} {metrics.get('skipped', 0)}")
-    print(f"{'-' * w}")
-    print(
-        f"  {'WER (corpus, micro-avg):':<{lw}} "
-        f"{metrics.get('wer_corpus', 0):.4f} "
-        f"({metrics.get('wer_corpus', 0) * 100:.2f}%)"
-    )
-    print(f"{'-' * w}")
-    print(
-        f"  {'WER per-sample mean:':<{lw}} "
-        f"{metrics.get('wer_per_sample_mean', 0):.4f} "
-        f"({metrics.get('wer_per_sample_mean', 0) * 100:.2f}%)"
-    )
-    print(
-        f"  {'WER per-sample median:':<{lw}} "
-        f"{metrics.get('wer_per_sample_median', 0):.4f}"
-    )
-    print(
-        f"  {'WER per-sample std:':<{lw}} "
-        f"{metrics.get('wer_per_sample_std', 0):.4f}"
-    )
-    print(
-        f"  {'WER per-sample p95:':<{lw}} "
-        f"{metrics.get('wer_per_sample_p95', 0):.4f}"
-    )
-    print(
-        f"  {'WER per-sample max:':<{lw}} "
-        f"{metrics.get('wer_per_sample_max', 0):.4f} "
-        f"({metrics.get('wer_per_sample_max', 0) * 100:.2f}%)"
-    )
-    print(
-        f"  {'WER corpus (excl >50%):':<{lw}} "
-        f"{metrics.get('wer_below_50_corpus', 0):.4f} "
-        f"({metrics.get('wer_below_50_corpus', 0) * 100:.2f}%)"
-    )
-    print(
-        f"  {'>50% WER samples:':<{lw}} "
-        f"{metrics.get('n_above_50_pct_wer', 0)} "
-        f"({metrics.get('pct_above_50_pct_wer', 0):.1f}%)"
-    )
-    print(f"{'-' * w}")
-    print(f"  {'Latency mean (s):':<{lw}} {metrics.get('latency_mean_s', 'N/A')}")
-    print(
-        f"  {'Audio duration mean (s):':<{lw}} "
-        f"{metrics.get('audio_duration_mean_s', 'N/A')}"
-    )
-    print(f"{'=' * w}\n")
-
-
-def calculate_asr_speed_metrics(outputs: list[SampleOutput]) -> dict:
-    """Compute speed metrics for the ASR transcription phase."""
-    successes = [o for o in outputs if o.is_success and o.asr_latency_s > 0]
-    if not successes:
-        return {
-            "total_samples": len(outputs),
-            "evaluated": 0,
-            "skipped": len(outputs),
-            "asr_latency_mean_s": 0.0,
-            "asr_latency_median_s": 0.0,
-            "asr_latency_p95_s": 0.0,
-            "asr_latency_p99_s": 0.0,
-            "asr_total_time_s": 0.0,
-            "asr_throughput_samples_per_s": 0.0,
-            "asr_rtf_mean": 0.0,
-            "asr_rtf_median": 0.0,
-            "asr_audio_processed_s": 0.0,
-        }
-
-    latencies = np.array([o.asr_latency_s for o in successes])
-    total_asr_time = float(np.sum(latencies))
-
-    audio_durations = [o.audio_duration_s for o in successes if o.audio_duration_s > 0]
-    rtfs = np.array(
-        [
-            o.asr_latency_s / o.audio_duration_s
-            for o in successes
-            if o.audio_duration_s > 0
-        ]
-    )
-
-    return {
-        "total_samples": len(outputs),
-        "evaluated": len(successes),
-        "skipped": len(outputs) - len(successes),
-        "asr_latency_mean_s": float(np.mean(latencies)),
-        "asr_latency_median_s": float(np.median(latencies)),
-        "asr_latency_p95_s": float(np.percentile(latencies, 95)),
-        "asr_latency_p99_s": float(np.percentile(latencies, 99)),
-        "asr_total_time_s": total_asr_time,
-        "asr_throughput_samples_per_s": (
-            float(len(successes) / total_asr_time) if total_asr_time > 0 else 0.0
-        ),
-        "asr_rtf_mean": float(np.mean(rtfs)) if len(rtfs) > 0 else 0.0,
-        "asr_rtf_median": float(np.median(rtfs)) if len(rtfs) > 0 else 0.0,
-        "asr_audio_processed_s": (
-            float(sum(audio_durations)) if audio_durations else 0.0
-        ),
-    }
-
-
-def print_asr_speed_summary(metrics: dict, model_name: str) -> None:
-    """Print ASR speed metrics summary table."""
-    lw = SUMMARY_LABEL_WIDTH
-    w = SUMMARY_LINE_WIDTH
-    print(f"\n{'=' * w}")
-    print(f"{'ASR Speed Benchmark Result':^{w}}")
-    print(f"{'=' * w}")
-    print(f"  {'Model:':<{lw}} {model_name}")
-    print(
-        f"  {'Evaluated / Total:':<{lw}} "
-        f"{metrics.get('evaluated', 0)}/{metrics.get('total_samples', 0)}"
-    )
-    print(f"  {'Skipped:':<{lw}} {metrics.get('skipped', 0)}")
-    print(f"{'-' * w}")
-    print(
-        f"  {'ASR latency mean (s):':<{lw}} "
-        f"{metrics.get('asr_latency_mean_s', 'N/A')}"
-    )
-    print(
-        f"  {'ASR latency median (s):':<{lw}} "
-        f"{metrics.get('asr_latency_median_s', 'N/A')}"
-    )
-    print(
-        f"  {'ASR latency p95 (s):':<{lw}} "
-        f"{metrics.get('asr_latency_p95_s', 'N/A')}"
-    )
-    print(
-        f"  {'ASR latency p99 (s):':<{lw}} "
-        f"{metrics.get('asr_latency_p99_s', 'N/A')}"
-    )
-    print(f"  {'ASR RTF mean:':<{lw}} {metrics.get('asr_rtf_mean', 'N/A')}")
-    print(f"  {'ASR RTF median:':<{lw}} {metrics.get('asr_rtf_median', 'N/A')}")
-    print(
-        f"  {'ASR total time (s):':<{lw}} " f"{metrics.get('asr_total_time_s', 'N/A')}"
-    )
-    print(
-        f"  {'ASR throughput (samples/s):':<{lw}} "
-        f"{metrics.get('asr_throughput_samples_per_s', 'N/A')}"
-    )
-    if metrics.get("asr_audio_processed_s"):
-        print(
-            f"  {'Audio processed (s):':<{lw}} " f"{metrics['asr_audio_processed_s']}"
-        )
-    print(f"{'=' * w}")
 
 
 def save_wer_results(
@@ -1239,82 +1015,6 @@ def make_tts_send_fn(
         return result
 
     return send_fn
-
-
-# ---------------------------------------------------------------------------
-# Speed results builders
-# ---------------------------------------------------------------------------
-
-
-def print_speed_summary(
-    metrics: dict,
-    model_name: str,
-    concurrency: int | None = None,
-    title: str = "TTS Benchmark Result",
-) -> None:
-    lw = SUMMARY_LABEL_WIDTH
-    w = SUMMARY_LINE_WIDTH
-    print(f"\n{'=' * w}")
-    print(f"{title:^{w}}")
-    print(f"{'=' * w}")
-    print(f"  {'Model:':<{lw}} {model_name}")
-    if concurrency is not None:
-        print(f"  {'Concurrency:':<{lw}} {concurrency}")
-    print(f"  {'Completed requests:':<{lw}} {metrics['completed_requests']}")
-    print(f"  {'Failed requests:':<{lw}} {metrics['failed_requests']}")
-    print(f"{'-' * w}")
-    print(f"  {'Latency mean (s):':<{lw}} {metrics.get('latency_mean_s', 'N/A')}")
-    print(f"  {'Latency median (s):':<{lw}} {metrics.get('latency_median_s', 'N/A')}")
-    print(f"  {'Latency p95 (s):':<{lw}} {metrics.get('latency_p95_s', 'N/A')}")
-    print(f"  {'Latency p99 (s):':<{lw}} {metrics.get('latency_p99_s', 'N/A')}")
-    if metrics.get("rtf_mean") is not None:
-        print(f"  {'RTF mean:':<{lw}} {metrics['rtf_mean']}")
-        print(f"  {'RTF median:':<{lw}} {metrics['rtf_median']}")
-    if metrics.get("audio_duration_mean_s"):
-        print(
-            f"  {'Audio duration mean (s):':<{lw}} {metrics['audio_duration_mean_s']}"
-        )
-    if metrics.get("tok_per_s_mean") is not None:
-        print(f"  {'Tok/s (per-req mean):':<{lw}} {metrics['tok_per_s_mean']}")
-        print(f"  {'Tok/s (per-req median):':<{lw}} {metrics['tok_per_s_median']}")
-    if metrics.get("tok_per_s_agg") is not None:
-        print(f"  {'Tok/s (aggregate):':<{lw}} {metrics['tok_per_s_agg']}")
-    if metrics.get("gen_tokens_mean") is not None:
-        print(f"  {'Gen tokens (mean):':<{lw}} {metrics['gen_tokens_mean']:.0f}")
-        print(f"  {'Gen tokens (total):':<{lw}} {metrics['gen_tokens_total']}")
-    if metrics.get("prompt_tokens_mean") is not None:
-        print(f"  {'Prompt tokens (mean):':<{lw}} {metrics['prompt_tokens_mean']:.0f}")
-        print(f"  {'Prompt tokens (total):':<{lw}} {metrics['prompt_tokens_total']}")
-    print(f"  {'Throughput (req/s):':<{lw}} {metrics.get('throughput_qps', 'N/A')}")
-    print(f"{'=' * w}")
-
-
-def build_speed_results(
-    outputs: list[RequestResult],
-    metrics: dict,
-    config: dict,
-) -> dict:
-    return {
-        "summary": metrics,
-        "config": config,
-        "per_request": [_request_result_to_dict(output) for output in outputs],
-    }
-
-
-def _request_result_to_dict(output: RequestResult) -> dict:
-    return {
-        "id": output.request_id,
-        "text": output.text,
-        "is_success": output.is_success,
-        "latency_s": round(output.latency_s, 4),
-        "audio_duration_s": round(output.audio_duration_s, 4),
-        "rtf": round(output.rtf, 4) if output.rtf < float("inf") else None,
-        "prompt_tokens": output.prompt_tokens or None,
-        "completion_tokens": output.completion_tokens or None,
-        "tok_per_s": round(output.tok_per_s, 1) if output.tok_per_s > 0 else None,
-        "wav_path": output.wav_path or None,
-        "error": output.error or None,
-    }
 
 
 def save_generated_audio_metadata(
