+++ tests/test_model/test_qwen3_omni_videomme_talker_ci.py
@@ -31,8 +31,9 @@ import pytest
 from benchmarks.dataset.prepare import DATASETS
 from benchmarks.dataset.videomme import VideoMMESample, load_videomme_samples
 from benchmarks.eval.benchmark_omni_videomme import VideoEvalConfig, run_video_eval
-from benchmarks.tasks.tts import print_speed_summary, print_wer_summary
-from benchmarks.tasks.video_understanding import print_videomme_accuracy_summary
+from benchmarks.metrics.performance import print_speed_summary
+from benchmarks.metrics.video import print_videomme_accuracy_summary
+from benchmarks.metrics.wer import print_wer_summary
 from tests.utils import (
     ServerHandle,
     apply_slack,
